# docs/swagger.yaml
basePath: /
definitions:
  github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      request_id:
        type: string
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta:
    properties:
      cursor:
        type: string
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta'
      success:
        type: boolean
    type: object
...
  internal_walletgroup.CreateGroupRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  internal_walletgroup.Group:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      wallet_count:
        type: integer
    type: object
  internal_walletgroup.UpdateGroupRequest:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@arbitrage.com
    name: API Support
  description: Backend API for arbitrage trading platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Arbitrage Platform API
  version: "1.0"
paths:
...
  /api/v1/auth/wallet/{wallet_id}:
    delete:
      description: Remove a wallet from the authenticated user
      parameters:
      - description: Wallet ID
        in: path
        name: wallet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
      security:
      - BearerAuth: []
      summary: Unlink wallet from account
      tags:
      - auth
      - web3
...
  /api/v1/groups:
    get:
      description: List the current user's wallet groups with wallet counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_walletgroup.Group'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List wallet groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a wallet group owned by the current user. Duplicate name
        within the same user is 409 GROUP-002.
      parameters:
      - description: Group to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_walletgroup.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_walletgroup.Group'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Create wallet group
      tags:
      - groups
  /api/v1/groups/{id}:
    delete:
      description: Delete an owned group; memberships are removed, wallets stay
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Delete wallet group
      tags:
      - groups
    get:
      description: Get one group with its wallet count
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_walletgroup.Group'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Get wallet group
      tags:
      - groups
    patch:
      consumes:
      - application/json
      description: Update name/description/color of an owned group. Renaming to an
        existing name of the same user is 409 GROUP-002.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_walletgroup.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_walletgroup.Group'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Update wallet group
      tags:
      - groups
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"